fix(rebase): resolve cli.rs refactor conflicts; re-apply surface-parity commands
Main extracted CliMode/GlobalFlags/run() to cli/run.rs after this branch diverged.
Re-apply surface-parity additions to the correct split files:
- cli/args.rs: add SourceIps/Timeline/Patterns/IngestRate/Sig/Notify variants,
  SigCommand/NotifyCommand enums, and all new Arg structs
- cli/run.rs: add dispatch arms for 6 new CliCommand variants + import SigCommand/NotifyCommand
- cli.rs: add parse routing, parse_source_ips/timeline/patterns/ingest_rate/sig/notify fns,
  parse_i64_flag helper, and FlagCursor::match_value method
