+++ src/cli/args.rs
@@ -18,6 +18,25 @@ pub(crate) enum CliCommand {
     Setup(SetupCommand),
     Db(DbCommand),
     Config(ConfigCommand),
+    SourceIps(SourceIpsArgs),
+    Timeline(TimelineArgs),
+    Patterns(PatternsArgs),
+    IngestRate(IngestRateArgs),
+    Sig(SigCommand),
+    Notify(NotifyCommand),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) enum SigCommand {
+    List(SigListArgs),
+    Ack(SigAckArgs),
+    Unack(SigUnackArgs),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) enum NotifyCommand {
+    Recent(NotifyRecentArgs),
+    Test(NotifyTestArgs),
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -397,38 +416,74 @@ pub(crate) struct AiPruneCheckpointsArgs {
 }
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub(crate) struct AiSimilarArgs {
-    pub query: String,
-    pub hostname: Option<String>,
-    pub app_name: Option<String>,
-    pub severity_min: Option<String>,
-    pub from: Option<String>,
-    pub to: Option<String>,
-    pub window_minutes: Option<u32>,
+pub(crate) struct SourceIpsArgs {
     pub limit: Option<u32>,
+    pub offset: Option<u32>,
     pub json: bool,
 }
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub(crate) struct AiAskHistoryArgs {
-    pub query: String,
-    pub hostname: Option<String>,
-    pub app_name: Option<String>,
+pub(crate) struct TimelineArgs {
+    pub bucket: Option<String>,
+    pub group_by: Option<String>,
     pub from: Option<String>,
     pub to: Option<String>,
-    pub limit: Option<u32>,
+    pub hostname: Option<String>,
+    pub app_name: Option<String>,
+    pub severity_min: Option<String>,
     pub json: bool,
 }
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub(crate) struct AiIncidentContextArgs {
-    pub from: String,
-    pub to: String,
+pub(crate) struct PatternsArgs {
+    pub from: Option<String>,
+    pub to: Option<String>,
     pub hostname: Option<String>,
     pub app_name: Option<String>,
-    pub query: Option<String>,
     pub severity_min: Option<String>,
+    pub scan_limit: Option<u32>,
+    pub top_n: Option<u32>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct IngestRateArgs {
+    pub by_host: bool,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct SigListArgs {
     pub limit: Option<u32>,
+    pub include_acknowledged: bool,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct SigAckArgs {
+    pub signature_hash: String,
+    pub notes: Option<String>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct SigUnackArgs {
+    pub signature_hash: String,
+    pub reason: Option<String>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct NotifyRecentArgs {
+    pub limit: Option<i64>,
+    pub rule_id: Option<String>,
+    pub since: Option<String>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct NotifyTestArgs {
+    pub body: Option<String>,
     pub json: bool,
 }
